Move Scrollable trait to kas-core
