+++ crates/kas-core/src/core/scroll_traits.rs
@@ -1,0 +1,45 @@
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License in the LICENSE-APACHE file or at:
+//     https://www.apache.org/licenses/LICENSE-2.0
+
+//! Scrollbar traits
+
+use super::Widget;
+use crate::event::EventMgr;
+use crate::geom::{Offset, Size};
+
+/// Additional functionality on scrollable widgets
+///
+/// This trait should be implemented by widgets supporting scrolling, enabling
+/// a parent to control scrolling.
+///
+/// If the widget scrolls itself it should set a scroll action via [`EventMgr::set_scroll`].
+pub trait Scrollable: Widget {
+    /// Given size `size`, returns whether `(horiz, vert)` scrolling is required
+    fn scroll_axes(&self, size: Size) -> (bool, bool);
+
+    /// Get the maximum scroll offset
+    ///
+    /// Note: the minimum scroll offset is always zero.
+    fn max_scroll_offset(&self) -> Offset;
+
+    /// Get the current scroll offset
+    ///
+    /// Contents of the scroll region are translated by this offset (to convert
+    /// coordinates from the outer region to the scroll region, add this offset).
+    ///
+    /// The offset is restricted between [`Offset::ZERO`] and
+    /// [`ScrollRegion::max_scroll_offset`].
+    fn scroll_offset(&self) -> Offset;
+
+    /// Set the scroll offset
+    ///
+    /// This may be used for programmatic scrolling, e.g. by a wrapping widget
+    /// with scroll controls. Note that calling this method directly on the
+    /// scrolling widget will not update any controls in a wrapping widget.
+    ///
+    /// The offset is clamped to the available scroll range and applied. The
+    /// resulting offset is returned.
+    fn set_scroll_offset(&mut self, mgr: &mut EventMgr, offset: Offset) -> Offset;
+}
