+++ crates/kas-core/src/prelude.rs
@@ -35,6 +35,8 @@ pub use crate::theme::{DrawMgr, SizeMgr, ThemeControl};
 #[doc(no_inline)]
 pub use crate::CoreData;
 #[doc(no_inline)]
+pub use crate::Scrollable;
+#[doc(no_inline)]
 pub use crate::TkAction;
 #[doc(no_inline)]
 pub use crate::WidgetId;
