+++ crates/kas-widgets/src/scroll_label.rs
@@ -5,7 +5,6 @@
 
 //! Scrollable and selectable label
 
-use super::Scrollable;
 use kas::event::components::{TextInput, TextInputAction};
 use kas::event::{Command, CursorIcon, Scroll, ScrollDelta};
 use kas::geom::Vec2;
